+++ src/main.ts
@@ -4,6 +4,8 @@ export { default as DSIconButton } from './components/DSIconButton'
 export { default as DSMenu } from './components/DSMenu'
 export { default as DSMenuLink } from './components/DSMenuLink'
 export { default as DSList } from './components/DSInput'
+export { default as DSModal } from './components/DSModal'
+export { default as DSSelect } from './components/DSSelect'
 
 export * as DateUtils from './utils/DateUtils'
 export * as StringUtils from './utils/StringUtils'