+++ src/types/active.d.ts
@@ -1,0 +1,1 @@
+type Active = 'on' | 'off'